+++ examples/app_shape_elements/app.tsx
@@ -18,6 +18,7 @@ type AppElementData = {
   paths: {
     d: string;
     fill: {
+      dropTarget: boolean;
       color: string;
     };
   }[];
@@ -39,6 +40,7 @@ const initialState: UIState = {
     {
       d: "M 0 0 H 100 V 100 H 0 L 0 0",
       fill: {
+        dropTarget: false,
         color: "#ff0099",
       },
     },
@@ -96,6 +98,7 @@ export const App = () => {
                           {
                             d: "",
                             fill: {
+                              dropTarget: false,
                               color: "#000000",
                             },
                           },